% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LymphomaSubtype-package.R
\docType{package}
\name{LymphomaSubtype}
\alias{LymphomaSubtype}
\title{Lymphoma Subtype Classification Data}
\description{
Use this package to load all the lymphoma subtypes as a tibble.
}
\author{
Makoto Kishida
}
